+++ src/assets/postMsg.js
@@ -1,0 +1,10 @@
+const vscode = acquireVsCodeApi();
+
+function postMessage(command, mimeType, text) {
+	var message = {
+		command: command,
+		mimeType: mimeType,
+		text: text,
+	};
+	vscode.postMessage(message);
+}