+++ src/utils/view.ts
@@ -1,8 +1,12 @@
-import * as vscode from 'vscode';
+import * as fs from 'fs';
 import * as path from 'path';
+import * as vscode from 'vscode';
 import { Base64Utils } from './base64utils';
+import { Localizer } from './localizer';
+import { MimeTypes } from './mimeTypes';
 
 export class View {
+	private messages: any;
 	private style = `
     body {
         background-color: #1e1e1e;
@@ -33,8 +37,31 @@ export class View {
         width: 100%;
     }
     
+    .action-button {
+        background-color: #303030;
+        border: #dddddd solid 1px;
+        color: #dddddd;
+        font-size: 1.1em;
+        margin: 8px 12px;
+        max-width: fit-content;
+        padding: 8px;
+    }
+
+    .button-bar {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+    }
+
+    .centered-column {
+        display: flex;
+        justify-content: center;
+    }
+    
     .content {
         border-top: #909090 solid 1px;
+        display: flex;
+        justify-content: center;
         margin: 4px;
         padding: 4px 0;
     }
@@ -44,12 +71,6 @@ export class View {
         display: flex;
         flex-direction: column;
         justify-content: center;
-    }
-    
-    .img-content {
-        border-top: #909090 solid 1px;
-        margin: 4px;
-        padding: 8px 0;
     }
     
     .page-content {
@@ -64,19 +85,16 @@ export class View {
         align-items: center;
     }
     
-    .page-nav > button {
-        margin: 0 12px;
-    }
-    
     .pdf-content {
         border-top: #909090 solid 1px;
         margin-top: 4px;
-        padding: 4px 0;
+        padding: 0;
     }
     
     .pdf-navbar {
         background-color: #454545;
         display: flex;
+        font-size: 1.1em;
         justify-content: space-between;
         padding: 8px;
     }
@@ -84,10 +102,10 @@ export class View {
     .pdf-navbar button {
         background-color: #303030;
         border: #dddddd solid 1px;
-        border-radius: 4px;
         color: #dddddd;
         font-weight: bold;
-        padding: 4px 8px;
+        padding: 6px 10px;
+        margin: 0 16px;
     }
     
     .spacer {
@@ -116,17 +134,12 @@ export class View {
     .two-col > :last-child {
         flex-grow: 1;
     }
-    
-    #switchButton {
-        background-color: #303030;
-        border: #ffffff solid 1px;
-        border-radius: 4px;
-        color: #ffffff;
-        max-width: fit-content;
-        margin: 4px;
-        padding: 4px;
-    }
     `;
+
+	constructor() {
+		let localizer = new Localizer();
+		this.messages = localizer.getLocalizedMessages();
+	}
 
 	public createView(
 		extensionRoot: vscode.Uri,
@@ -136,10 +149,31 @@ export class View {
 		filePath?: string,
 	) {
 		// Create and show panel
-		var webviewPanel = vscode.window.createWebviewPanel('base64viewer', 'Base64 Viewer', vscode.ViewColumn.Two, {});
+		var webviewPanel = vscode.window.createWebviewPanel(
+			'base64viewer',
+			this.messages.general.title,
+			vscode.ViewColumn.Two,
+			{},
+		);
 		webviewPanel.webview.options = {
 			enableScripts: true,
 		};
+
+		// Clean resources
+		webviewPanel.onDidDispose(() => {
+			webviewPanel.dispose();
+		});
+
+		// Handle messages from the webview
+		webviewPanel.webview.onDidReceiveMessage((message) => {
+			if (message.command === 'copy') {
+				this.copyToClipboard(message.text);
+				return;
+			} else if (message.command === 'save') {
+				this.saveDecodedFile(message.mimeType, message.text);
+				return;
+			}
+		});
 
 		if (viewType === 'decoding') {
 			webviewPanel.webview.html = this.initWebviewDecodingContent(
@@ -159,6 +193,21 @@ export class View {
 		}
 	}
 
+	private copyToClipboard(text: string) {
+		vscode.env.clipboard.writeText(text).then(() => {
+			this.showInformationPopup(this.messages.general.copiedToClipboard);
+		});
+	}
+
+	private getNonce() {
+		let text = '';
+		const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+		for (let i = 0; i < 32; i++) {
+			text += possible.charAt(Math.floor(Math.random() * possible.length));
+		}
+		return text;
+	}
+
 	private initWebviewDecodingContent(
 		extensionRoot: vscode.Uri,
 		webview: vscode.Webview,
@@ -166,6 +215,7 @@ export class View {
 		mimeType: string,
 	): string {
 		const b64u = new Base64Utils();
+		const nonce = this.getNonce();
 		const spacer = '  |  ';
 		const resolveAsUri = (...p: string[]): vscode.Uri => {
 			const uri = vscode.Uri.file(path.join(extensionRoot.path, ...p));
@@ -185,27 +235,34 @@ export class View {
                         <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
                         <meta name="google" content="notranslate">
                         <meta http-equiv="X-UA-Compatible" content="IE=edge">
-                        <title>Base 64 Viewer</title>
-                        <script src="${resolveAsUri('lib', 'pdfjs-dist', 'pdf.js')}"></script>
+                        <title>${this.messages.general.title}</title>
+                        <script nonce="${nonce}" src="${resolveAsUri('lib', 'pdfjs-dist', 'pdf.js')}"></script>
                         <style>${this.style}</style>
                     </head>`;
 			body = `
                 <body>
                     <div class="title-bar">
-                        <h1>Base 64 Viewer</h1>
+                        <h1>${this.messages.general.title}</h1>
                     </div>
             
                     <div class="page-content two-col">
                         <div>
                             <h3>${mimeType}  (${fileSize})</h3>
                             <div class="pdf-content">
+                                <div class="button-bar">
+                                    <button class="action-button" onclick="postMessage('save', '${mimeType}', '${base64String}')">
+                                        ${this.messages.general.saveButton}
+                                    </button>
+                                </div>
                                 <div class="pdf-navbar">
                                     <div class="spacer"></div>
             
                                     <div class="page-nav">
                                         <button onclick="changePage(loadedPdf, currentPage, 'prev')"><</button>
                                         <span>
-                                            Page <span id="currentPage"></span> / <span id="totalPage"></span>
+                                            ${
+												this.messages.pdf.page
+											} : <span id="currentPage"></span> / <span id="totalPage"></span>
                                         </span>
                                         <button onclick="changePage(loadedPdf, currentPage, 'next')">></button>
                                     </div>
@@ -218,14 +275,22 @@ export class View {
                         </div>
             
                         <div>
-                            <h3>Ordered PDF Elements</h3>
-                            <div class="content">
-                                <code id="pdfElementList"></code>
+                            <div>
+                                <h3>${this.messages.pdf.orderedElements.text.title}</h3>
+                                <div class="content">
+                                    <code id="pdfTextElementsList"></code>
+                                </div>
+                            </div>
+                            <div>
+                                <h3>${this.messages.pdf.orderedElements.images.title}</h3>
+                                <div class="content" id="pdfImagesList"></div>
                             </div>
                         </div>
                     </div>
             
-                    <script>
+                    <script nonce="${nonce}" src="${resolveAsUri('src', 'assets', 'postMsg.js')}"></script>
+            
+                    <script nonce="${nonce}">
                         var pdfData = atob('${base64String}');
                         var pdfjsLib = window['pdfjs-dist/build/pdf'];
                         pdfjsLib.GlobalWorkerOptions.workerSrc = '${resolveAsUri(
@@ -254,7 +319,7 @@ export class View {
                             renderPage(pdf, pageNumber);
                         
                             // Parsing the pdf page by page
-                            parsePdf(pdf);                            
+                            parsePdf(pdf);
                         }, function (reason) {
                             // PDF loading error
                             console.error("Error: " + reason);
@@ -272,26 +337,54 @@ export class View {
                             renderPage(pdf, pageNumber);
                         }
                         
+                        function extractImagesInPage(page) {
+                            const scale = 1.5;
+                            const viewport = page.getViewport({scale: scale});
+
+                            page.getOperatorList().then(function(opList) {
+                                var svgGfx = new pdfjsLib.SVGGraphics(page.commonObjs, page.objs);
+                                return svgGfx.getSVG(opList, viewport);
+                            }).then(function(svg) {
+                                var pageSvgString = new XMLSerializer().serializeToString(svg);
+                                var cutSvg = pageSvgString.split('<svg:image ');
+                                for (var i=0; i < cutSvg.length; i++) {
+                                    var recut = cutSvg[i+1].split('/>');
+                                    var blob = recut[0].split('href="');
+                                    blob = blob[1].split('"');
+                                    var svgSrc = blob[0];
+                                    
+                                    var img = document.createElement("IMG");
+                                    img.setAttribute('src', svgSrc)
+                                    img.setAttribute('width', '80%');
+
+                                    document.getElementById('pdfImagesList').appendChild(img);
+                                }
+                            });	
+                        }
+                        
+                        function extractTextInPage(page, htmlList) {
+                            page.getTextContent().then(function(tokenizedText) {
+                                var textElementsList = "";
+                                var pageContent = tokenizedText.items.map(token => token.str);
+                            
+                                pageContent.forEach(function(textElement) {
+                                    textElement = textElement.trim();
+                                
+                                    if (textElement !== '') {
+                                        textElementsList = textElementsList + textElement + '${spacer}';
+                                    }
+                                });
+
+                                htmlList.innerText = htmlList.innerText + textElementsList;
+                            });
+                        }
+                        
                         function parsePdf(pdf) {
-                            var fileElementList = "";
-                            var list = document.getElementById('pdfElementList');
+                            var pdfTextElementsList = document.getElementById('pdfTextElementsList');
                             for (let i = 0; i < pdf.numPages; i++) {                            
                                 pdf.getPage(i + 1).then(function(page) {
-                                    let pageContent;
-                                
-                                    page.getTextContent().then(function(tokenizedText) {
-                                        pageContent = tokenizedText.items.map(token => token.str);
-                                    
-                                        pageContent.forEach(function(textElement) {
-                                            textElement = textElement.trim();
-                                        
-                                            if (textElement !== '') {
-                                                fileElementList = fileElementList + textElement + '${spacer}';
-                                            }
-                                        });
-                                    
-                                        list.innerText = fileElementList;
-                                    });
+                                    extractTextInPage(page, pdfTextElementsList);
+                                    extractImagesInPage(page);
                                 });				
                             }
                         }
@@ -333,21 +426,28 @@ export class View {
                         <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
                         <meta name="google" content="notranslate">
                         <meta http-equiv="X-UA-Compatible" content="IE=edge">
-                        <title>Base 64 Viewer</title>
+                        <title>${this.messages.general.title}</title>
                         <style>${this.style}</style>
                     </head>`;
 			body = `
                 <body>
                     <div class="title-bar">
-                        <h1>Base 64 Viewer</h1>
+                        <h1>${this.messages.general.title}</h1>
                     </div>
             
                     <div class="page-content">
                         <h3>${mimeType}  (${fileSize})</h3>
-                        <div class="img-content">
+                        <div class="content">
+                            <div class="button-bar">
+                                <button class="action-button" onclick="postMessage('save', '${mimeType}', '${base64String}')">
+                                    ${this.messages.general.saveButton}
+                                </button>
+                            </div>
                             <img src="data:${mimeType};base64,${base64String}"/>
                         </div>
                     </div>
+            
+                    <script nonce="${nonce}" src="${resolveAsUri('src', 'assets', 'postMsg.js')}"></script>
                 </body>
             </html>`;
 		} else if (mimeType.includes('text')) {
@@ -359,23 +459,30 @@ export class View {
                         <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
                         <meta name="google" content="notranslate">
                         <meta http-equiv="X-UA-Compatible" content="IE=edge">
-                        <title>Base 64 Viewer</title>
+                        <title>${this.messages.general.title}</title>
                         <style>${this.style}</style>
                     </head>`;
 			body = `
                 <body>
                     <div class="title-bar">
-                        <h1>Base 64 Viewer</h1>
+                        <h1>${this.messages.general.title}</h1>
                     </div>
             
                     <div class="page-content">
                         <h3>${mimeType}  (${fileSize})</h3>
                         <div class="content">
+                            <div class="button-bar">
+                                <button class="action-button" onclick="postMessage('save', '${mimeType}', '${base64String}')">
+                                    ${this.messages.general.saveButton}
+                                </button>
+                            </div>
                             <code id="code-tag"></code>
                         </div>
                     </div>
+            
+                    <script nonce="${nonce}" src="${resolveAsUri('src', 'assets', 'postMsg.js')}"></script>
                         
-                    <script>
+                    <script nonce="${nonce}">
                         var text = atob('${base64String}');
                         var codeTag = document.getElementById('code-tag');
                         codeTag.innerText = text;
@@ -391,21 +498,28 @@ export class View {
                         <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
                         <meta name="google" content="notranslate">
                         <meta http-equiv="X-UA-Compatible" content="IE=edge">
-                        <title>Base 64 Viewer</title>
+                        <title>${this.messages.general.title}</title>
                         <style>${this.style}</style>
                     </head>`;
 			body = `
                 <body>
                     <div class="title-bar">
-                        <h1>Base 64 Viewer</h1>
+                        <h1>${this.messages.general.title}</h1>
                     </div>
             
                     <div class="page-content">
                         <h3>${mimeType}  (${fileSize})</h3>
                         <div class="content">
-                            <h2>This format can't be displayed!</h2>
+                            <div class="button-bar">
+                                <button class="action-button" onclick="postMessage('save', '${mimeType}', '${base64String}')">
+                                    ${this.messages.general.saveButton}
+                                </button>
+                            </div>
+                            <h2>${this.messages.general.cantDisplayContent}</h2>
                         </div>
                     </div>
+            
+                    <script nonce="${nonce}" src="${resolveAsUri('src', 'assets', 'postMsg.js')}"></script>
                 </body>
             </html>`;
 		}
@@ -420,6 +534,7 @@ export class View {
 		mimeType: string,
 		filePath: string,
 	): string {
+		const nonce = this.getNonce();
 		const spacer = '  |  ';
 		const resolveAsUri = (...p: string[]): vscode.Uri => {
 			const uri = vscode.Uri.file(path.join(extensionRoot.path, ...p));
@@ -438,27 +553,40 @@ export class View {
                         <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
                         <meta name="google" content="notranslate">
                         <meta http-equiv="X-UA-Compatible" content="IE=edge">
-                        <title>Base 64 Viewer</title>
+                        <title>${this.messages.general.title}</title>
                         <style>${this.style}</style>
-                        <script src="${resolveAsUri('lib', 'pdfjs-dist', 'pdf.js')}"></script>
+                        <script nonce="${nonce}" src="${resolveAsUri('lib', 'pdfjs-dist', 'pdf.js')}"></script>
                     </head>`;
 			body = `
                 <body>
                     <div class="title-bar">
-                        <h1>Base 64 Viewer</h1>
+                        <h1>${this.messages.general.title}</h1>
                     </div>
             
                     <div class="page-content">
-                        <h3>${mimeType} ${filePath}</h3>
+                        <h3>${filePath}<br/><br/>${mimeType}</h3>
                         <div class="content encoded-content">
-                            <button id="switchButton" onclick="switchContent()">Switch to Ordered PDF Elements</button>
+                            <div class="button-bar">
+                                <button id="switchButton" class="action-button" onclick="switchContent()">${
+									this.messages.pdf.orderedElements.text.button
+								}</button>
+                                <button class="action-button" onclick="postMessage('copy', '${mimeType}', '${content}')">${
+				this.messages.general.copyButton
+			}</button>
+                            </div>
                             <code id="code-tag">${content}</code>
+                            <br/>
+                            <details id="pdfImagesList">
+                                <summary>${this.messages.pdf.orderedElements.images.title}</summary>
+                            </details>
                         </div>
                     </div>
+            
+                    <script nonce="${nonce}" src="${resolveAsUri('src', 'assets', 'postMsg.js')}"></script>
 
-                    <script>
+                    <script nonce="${nonce}">
                         var displayed = "content";
-                        var elementList = "";
+                        var textElementsList = "";
 
                         var pdfData = atob('${content}');
                         var pdfjsLib = window['pdfjs-dist/build/pdf'];
@@ -476,26 +604,52 @@ export class View {
                             // PDF loading error
                             console.error("Error: " + reason);
                         });
+                        
+                        function extractImagesInPage(page) {
+                            const scale = 1.5;
+                            const viewport = page.getViewport({scale: scale});
 
+                            page.getOperatorList().then(function(opList) {
+                                var svgGfx = new pdfjsLib.SVGGraphics(page.commonObjs, page.objs);
+                                return svgGfx.getSVG(opList, viewport);
+                            }).then(function(svg) {
+                                var pageSvgString = new XMLSerializer().serializeToString(svg);
+                                var cutSvg = pageSvgString.split('<svg:image ');
+                                for (var i=0; i < cutSvg.length; i++) {
+                                    var recut = cutSvg[i+1].split('/>');
+                                    var blob = recut[0].split('href="');
+                                    blob = blob[1].split('"');
+                                    var svgSrc = blob[0];
+
+                                    var img = document.createElement("IMG");
+                                    img.setAttribute('src', svgSrc)
+                                    img.setAttribute('width', '80%');
+
+                                    document.getElementById('pdfImagesList').appendChild(img);
+                                }
+                            });	
+                        }
+                        
+                        function extractTextInPage(page, htmlList) {
+                            page.getTextContent().then(function(tokenizedText) {
+                                var pageContent = tokenizedText.items.map(token => token.str);
+                            
+                                pageContent.forEach(function(textElement) {
+                                    textElement = textElement.trim();
+                                
+                                    if (textElement !== '') {
+                                        textElementsList = textElementsList + textElement + '${spacer}';
+                                    }
+                                });
+                            });
+                        }
+                        
                         function parsePdf(pdf) {
-                            var fileElementList = "";
+                            var pdfTextElementsList = document.getElementById('pdfTextElementsList');
                             for (let i = 0; i < pdf.numPages; i++) {                            
                                 pdf.getPage(i + 1).then(function(page) {
-                                    let pageContent;
-                                
-                                    page.getTextContent().then(function(tokenizedText) {
-                                        pageContent = tokenizedText.items.map(token => token.str);
-                                    
-                                        pageContent.forEach(function(textElement) {
-                                            textElement = textElement.trim();
-                                        
-                                            if (textElement !== '') {
-                                                fileElementList = fileElementList + textElement + '${spacer}';
-                                            }
-                                        });
-                                    
-                                        elementList = fileElementList;
-                                    });					
+                                    extractTextInPage(page, pdfTextElementsList);
+                                    extractImagesInPage(page);
                                 });				
                             }
                         }
@@ -505,13 +659,13 @@ export class View {
                             var switchButton = document.getElementById('switchButton');
 
                             if (displayed === "content") {
-                                codeTag.innerText = elementList;
-                                displayed = "elementList";
-                                switchButton.innerText = "Switch to Base64 Encoded File";
+                                codeTag.innerText = textElementsList;
+                                displayed = "textElementsList";
+                                switchButton.innerText = "${this.messages.pdf.encodedString.button}";
                             } else {
                                 codeTag.innerText = '${content}';
                                 displayed = "content";
-                                switchButton.innerText = "Switch to Ordered PDF Elements";
+                                switchButton.innerText = "${this.messages.pdf.orderedElements.text.button}";
                             }
                         }
                     </script>
@@ -526,25 +680,77 @@ export class View {
                         <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
                         <meta name="google" content="notranslate">
                         <meta http-equiv="X-UA-Compatible" content="IE=edge">
-                        <title>Base 64 Viewer</title>
+                        <title>${this.messages.general.title}</title>
                         <style>${this.style}</style>
                     </head>`;
 			body = `
                 <body>
                     <div class="title-bar">
-                        <h1>Base 64 Viewer</h1>
+                        <h1>${this.messages.general.title}</h1>
                     </div>
             
                     <div class="page-content">
-                        <h3>${mimeType} ${filePath}</h3>
-                        <div class="content">
+                        <h3>${filePath}<br/><br/>${mimeType}</h3>
+                        <div class="content encoded-content">
+                            <div class="button-bar">
+                                <button class="action-button" onclick="postMessage('copy', '${mimeType}', '${content}')">${
+				this.messages.general.copyButton
+			}</button>
+                            </div>
                             <code id="code-tag">${content}</code>
                         </div>
                     </div>
+            
+                    <script nonce="${nonce}" src="${resolveAsUri('src', 'assets', 'postMsg.js')}"></script>
                 </body>
             </html>`;
 		}
 
 		return head + body;
+	}
+
+	private saveDecodedFile(mimeType: string, base64String: string) {
+		const mT = new MimeTypes();
+		const xss = require('xss');
+
+		let mTKeys = [...mT.mimeTypes.entries()].filter(({ 1: v }) => v === mimeType).map(([k]) => k);
+		let fileExtension = mTKeys[0];
+
+		vscode.window
+			.showInputBox({
+				prompt: this.messages.general.prompt.save,
+			})
+			.then(
+				(name) => {
+					let fileName = '';
+
+					if (name !== undefined) {
+						fileName = xss(name);
+					} else {
+						fileName = new Date().toISOString();
+					}
+
+					fileName = fileName + '.' + fileExtension;
+
+					let buf = Buffer.from("'" + base64String + "'", 'base64');
+
+					fs.writeFile(fileName, buf, (err) => {
+						if (err) {
+							this.showErrorPopup(this.messages.general.fileSave.error + ' : ' + err);
+						} else {
+							this.showInformationPopup(this.messages.general.fileSave.success + ' : ' + fileName);
+						}
+					});
+				},
+				(reason) => this.showErrorPopup(reason),
+			);
+	}
+
+	private showErrorPopup(message: string) {
+		vscode.window.showErrorMessage(message);
+	}
+
+	private showInformationPopup(message: string) {
+		vscode.window.showInformationMessage(message);
 	}
 }